+++ packages/yourloops/pages/hcp/patients/utils.ts
@@ -29,9 +29,11 @@
 import moment from "moment-timezone"; // TODO: Change moment-timezone lib with something else
 import { TFunction } from "i18next";
 
-import { SortFields } from "../../../models/generic";
+import { PatientTableSortFields, SortFields } from "../../../models/generic";
 import { MedicalData } from "../../../models/device-data";
 import { MedicalTableValues } from "./models";
+import { TeamMember, TeamUser } from "../../../lib/team";
+import { Patient, PatientTeam } from "../../../models/patient";
 
 export const getMedicalValues = (medicalData: MedicalData | null | undefined, na = "N/A"): MedicalTableValues => {
   let tir = "-";
@@ -100,4 +102,107 @@ export const translateSortField = (t: TFunction, field: SortFields): string => {
     break;
   }
   return trOrderBy;
+};
+
+export const compareNumbers = (a: number, b: number): number => {
+  return a - b;
+};
+
+export const compareString = (a: string, b: string): number => {
+  return a.localeCompare(b);
+};
+
+export const compareDate = (a: Date, b: Date): number => {
+  return a.getTime() - b.getTime();
+};
+
+function compareValues(
+  a: string | number | Date | boolean | null | undefined,
+  b: string | number | boolean | Date | null | undefined
+) {
+  if (typeof a === "string" && typeof b === "string") {
+    return compareString(a, b);
+  }
+  if (a instanceof Date && b instanceof Date) {
+    return compareDate(a, b);
+  }
+  if (typeof a === "number" && typeof b === "number") {
+    return compareNumbers(a, b);
+  }
+  if (!a && b) {
+    return 1;
+  }
+  if (!b && a) {
+    return -1;
+  }
+  return 0;
+}
+
+/**
+ * Compare two patient for sorting the patient table
+ * @param a A patient
+ * @param b A patient
+ * @param orderBy Sort field
+ */
+export const comparePatients = (a: Patient, b: Patient, orderBy: PatientTableSortFields): number => {
+  let aValue: string | number | Date | boolean | null | undefined;
+  let bValue: string | number | Date | boolean | null | undefined;
+
+  switch (orderBy) {
+  case PatientTableSortFields.alertTimeTarget:
+    aValue = a.alerts?.timeSpentAwayFromTargetRate;
+    bValue = b.alerts?.timeSpentAwayFromTargetRate;
+    break;
+  case PatientTableSortFields.alertHypoglycemic:
+    aValue = a.alerts?.frequencyOfSevereHypoglycemiaRate;
+    bValue = b.alerts?.frequencyOfSevereHypoglycemiaRate;
+    break;
+  case PatientTableSortFields.flag:
+    aValue = a.flagged;
+    bValue = b.flagged;
+    break;
+  case PatientTableSortFields.ldu:
+    aValue = getMedicalValues(a.medicalData).lastUploadEpoch;
+    bValue = getMedicalValues(b.medicalData).lastUploadEpoch;
+    break;
+  case PatientTableSortFields.patientFullName:
+    aValue = a.fullName;
+    bValue = b.fullName;
+    break;
+  case PatientTableSortFields.remoteMonitoring:
+    aValue = a.remoteMonitoring;
+    bValue = b.remoteMonitoring;
+    break;
+  case PatientTableSortFields.system:
+    aValue = a.system;
+    bValue = b.system;
+    break;
+  }
+  return compareValues(aValue, bValue);
+};
+
+export const mapTeamMemberToPatientTeam = (member: TeamMember): PatientTeam => {
+  return {
+    code: member.team.code,
+    invitation: member.invitation,
+    status: member.status,
+    teamId: member.team.id,
+    teamName: member.team.name,
+  };
+};
+
+export const mapTeamUserToPatient = (teamUser: TeamUser): Patient => {
+  return {
+    alerts: null,
+    firstName: teamUser.profile?.firstName,
+    flagged: undefined,
+    fullName: teamUser.profile?.fullName ?? teamUser.username,
+    lastName: teamUser.profile?.lastName,
+    medicalData: null,
+    remoteMonitoring: undefined,
+    system: undefined,
+    teams: teamUser.members.map(member => mapTeamMemberToPatientTeam(member)),
+    userid: teamUser.userid,
+    username: teamUser.username,
+  };
 };