+++ packages/yourloops/test/pages/hcp/patients-table.test.tsx
@@ -30,18 +30,24 @@ import React from "react";
 import { render, unmountComponentAtNode } from "react-dom";
 import { act, Simulate } from "react-dom/test-utils";
 
-import { SortDirection, SortFields } from "../../../models/generic";
-import { useTeam, TeamContextProvider } from "../../../lib/team";
+import { PatientTableSortFields, SortDirection, UserInvitationStatus } from "../../../models/generic";
+import { TeamContextProvider, useTeam } from "../../../lib/team";
 import { NotificationContextProvider } from "../../../lib/notifications";
 import { AuthContext, AuthContextProvider } from "../../../lib/auth";
 
-import { stubNotificationContextValue } from "../../lib/notifications/hook.test";
 import "../../intersectionObserverMock";
 
-import PatientListTable from "../../../pages/hcp/patients/table";
+import PatientTable from "../../../pages/hcp/patients/table";
 import { teamAPI } from "../../lib/team/utils";
 import { loggedInUsers } from "../../common";
 import { createAuthHookStubs } from "../../lib/auth/utils";
+import { stubNotificationContextValue } from "../../lib/notifications/utils";
+import { TablePagination, ThemeProvider } from "@material-ui/core";
+import { getTheme } from "../../../components/theme";
+import renderer from "react-test-renderer";
+import PatientRow from "../../../pages/hcp/patients/row";
+import { createPatient, createPatientTeam } from "../../common/utils";
+import { Patient } from "../../../models/patient";
 
 describe("Patient list table", () => {
   const authHcp = loggedInUsers.hcpSession;
@@ -49,6 +55,20 @@ describe("Patient list table", () => {
   const clickPatientStub = jest.fn();
   const clickFlagPatientStub = jest.fn();
   const clickRemovePatientStub = jest.fn();
+
+  const team1Id = "team1Id";
+  const patient1 = createPatient("id1", [createPatientTeam(team1Id, UserInvitationStatus.accepted)]);
+  const patient2 = createPatient("id2", [createPatientTeam(team1Id, UserInvitationStatus.accepted)]);
+  const patient3 = createPatient("id3", [createPatientTeam(team1Id, UserInvitationStatus.accepted)]);
+  const patient4 = createPatient("id4", [createPatientTeam(team1Id, UserInvitationStatus.accepted)]);
+  const patient5 = createPatient("id5", [createPatientTeam(team1Id, UserInvitationStatus.accepted)]);
+  const patient6 = createPatient("id6", [createPatientTeam(team1Id, UserInvitationStatus.accepted)]);
+  const patient7 = createPatient("id7", [createPatientTeam(team1Id, UserInvitationStatus.accepted)]);
+  const patient8 = createPatient("id8", [createPatientTeam(team1Id, UserInvitationStatus.accepted)]);
+  const patient9 = createPatient("id9", [createPatientTeam(team1Id, UserInvitationStatus.accepted)]);
+  const patient10 = createPatient("id10", [createPatientTeam(team1Id, UserInvitationStatus.accepted)]);
+  const patient11 = createPatient("id11", [createPatientTeam(team1Id, UserInvitationStatus.accepted)]);
+  const allPatients = [patient1, patient2, patient3, patient4, patient5, patient6, patient7, patient8, patient9, patient10, patient11];
 
   let container: HTMLElement | null = null;
 
@@ -73,19 +93,19 @@ describe("Patient list table", () => {
     const patients = team.getPatients();
 
     return (
-      <PatientListTable
-        patients={patients}
-        flagged={[]}
-        order={SortDirection.asc}
-        orderBy={SortFields.lastname}
-        onClickPatient={clickPatientStub}
-        onFlagPatient={clickFlagPatientStub}
-        onSortList={jest.fn()}
-        onClickRemovePatient={clickRemovePatientStub}
-      />
+      <ThemeProvider theme={getTheme()}>
+        <PatientTable
+          patients={patients}
+          flagged={[]}
+          order={SortDirection.asc}
+          orderBy={PatientTableSortFields.patientFullName}
+          onClickPatient={clickPatientStub}
+          onFlagPatient={clickFlagPatientStub}
+          onSortList={jest.fn()}
+        />
+      </ThemeProvider>
     );
   };
-
 
   async function mountComponent(): Promise<void> {
     await act(() => {
@@ -100,6 +120,26 @@ describe("Patient list table", () => {
           </AuthContextProvider>, container, resolve);
       });
     });
+  }
+
+  function renderPatientList(teamUsers: Patient[]) {
+    return renderer.create(
+      <ThemeProvider theme={getTheme()}>
+        <AuthContextProvider value={authHookHcp}>
+          <TeamContextProvider teamAPI={teamAPI}>
+            <PatientTable
+              patients={teamUsers}
+              flagged={[]}
+              order={SortDirection.asc}
+              orderBy={PatientTableSortFields.patientFullName}
+              onClickPatient={clickPatientStub}
+              onFlagPatient={clickFlagPatientStub}
+              onSortList={jest.fn()}
+            />
+          </TeamContextProvider>
+        </AuthContextProvider>
+      </ThemeProvider>
+    );
   }
 
   it("should be able to render", async () => {
@@ -129,17 +169,18 @@ describe("Patient list table", () => {
     expect(clickFlagPatientStub).toHaveBeenCalledTimes(1);
   });
 
-  /*
-  * TODO Can't add this feature for the moment
-  *  we need to wait until yourloops will be certified to level 2 of medical device
-  *  see YLP-370 (https://diabeloop.atlassian.net/browse/YLP-370)
-   */
-  // it("should call onRemovePatient method when clicking on a remove icon", async () => {
-  //   await mountComponent();
-  //   const firstRow = document.querySelector(".patients-list-row");
-  //   const removeButton = firstRow.querySelector(".remove-patient-hcp-view-button");
-  //   Simulate.click(removeButton);
-  //   expect(clickRemovePatientStub.calledOnce).toBeTruthy();
-  // });
+  it("should display only 10 patients when number pagination is by 10", () => {
+    const component = renderPatientList(allPatients);
+    const patientRows = component.root.findAllByType(PatientRow);
+    expect(patientRows).toHaveLength(10);
+  });
+
+  it("should display all patients when number pagination is by 100", () => {
+    const component = renderPatientList(allPatients);
+    const tablePagination = component.root.findByType(TablePagination);
+    tablePagination.props.onRowsPerPageChange({ target: { value: "100" } });
+    const patientRows = component.root.findAllByType(PatientRow);
+    expect(patientRows).toHaveLength(allPatients.length);
+  });
 });